Close consent health race windows
