+++ tests/test_connection_service.py
@@ -3,24 +3,17 @@
 import asyncio
 import threading
 import time
-from collections.abc import Callable, Mapping
-from types import SimpleNamespace
-from typing import Any, cast
+from typing import Any
 
 import pytest
-from mcp.server.mcpserver import Context
 from mcp.server.mcpserver import MCPServer
 from mcp.server.mcpserver.exceptions import ToolError
-from mcp_types import CLIENT_INFO_META_KEY, Implementation
 
 from delta_exchange_mcp import credentials as credential_check
 from delta_exchange_mcp import setup, store
 from delta_exchange_mcp.auth import connection as connection_mod
 from delta_exchange_mcp.auth.connection import ConnectionService
 from delta_exchange_mcp.auth.consent import (
-    ConsentBackend,
-    ConsentBinding,
-    ConsentRevocationError,
     ConsentStorageError,
     ConsentStore,
     MemoryConsentBackend,
@@ -30,111 +23,19 @@ from delta_exchange_mcp.auth.store import (
     CredentialSource,
     CredentialState,
     CredentialStore,
-    CredentialStoreError,
     MetadataError,
     MemoryMetadata,
     MemorySecretBackend,
 )
 from delta_exchange_mcp.server import build_server
 from delta_exchange_mcp.tools import trading
-
-
-def context(name: str, version: str = "1") -> Context:
-    request = SimpleNamespace(
-        meta={
-            CLIENT_INFO_META_KEY: Implementation(name=name, version=version),
-        },
-        protocol_version="2026-07-28",
-    )
-    return Context(request_context=cast(Any, request))
-
-
-def stores(*, persistent: bool = True) -> tuple[CredentialStore, ConsentStore]:
-    source = CredentialSource.OS_STORE if persistent else CredentialSource.MEMORY
-    credentials = CredentialStore(
-        MemorySecretBackend(),
-        MemoryMetadata(),
-        source,
-    )
-    consent = ConsentStore(
-        store.path().with_name("consent.json"),
-        secure_backend_available=persistent,
-        memory_backend=MemoryConsentBackend(),
-    )
-    return credentials, consent
-
-
-def service(
-    validator: Any,
-    *,
-    persistent: bool = True,
-) -> ConnectionService:
-    credentials, consent = stores(persistent=persistent)
-    return ConnectionService.open(
-        credentials=credentials,
-        consent=consent,
-        validator=validator,
-    )
-
-
-def action(
-    connection: ConnectionService,
-    client_name: str,
-    name: str,
-    arguments: Mapping[str, Any],
-    revision: setup.Revision | None = None,
-) -> setup.ActionResult:
-    expected = connection._revision(client_name) if revision is None else revision
-    return connection._actions(client_name)(name, arguments, expected)
-
-
-async def verified(
-    environment: str, api_key: str, api_secret: str
-) -> credential_check.Check:
-    return credential_check.Check(
-        ok=True,
-        reachable=True,
-        detail="42",
-    )
-
-
-def _assert_place_order_blocked(
-    monkeypatch: pytest.MonkeyPatch,
-    connection: ConnectionService,
-    final_check: Callable[[], bool],
-) -> None:
-    gate = trading.TradeGate()
-    mcp = MCPServer("blocked-place-order")
-    trading.register(mcp, connection.client, None, gate)
-    mutations: list[str] = []
-
-    async def post(
-        path: str,
-        payload: dict[str, Any],
-        *,
-        auth: bool = False,
-    ) -> dict[str, Any]:
-        del payload, auth
-        mutations.append(path)
-        return {}
-
-    monkeypatch.setattr(connection.client, "post", post)
-
-    async def invoke() -> object:
-        gate.bind_final_check(final_check)
-        return await mcp.call_tool(
-            "place_order",
-            {
-                "product_id": 27,
-                "size": 1,
-                "side": "buy",
-                "order_type": "market_order",
-            },
-        )
-
-    with pytest.raises(ToolError, match="trading was disabled"):
-        asyncio.run(invoke())
-    assert mutations == []
+from tests.connection_support import (
+    action,
+    context,
+    service,
+    stores,
+    verified,
+)
 
 
 @pytest.mark.parametrize(
@@ -769,748 +670,6 @@ def test_status_and_results_never_contain_credential_material() -> None:
 
     assert "visible-only-to-store-key" not in rendered
     assert "visible-only-to-store-secret" not in rendered
-
-
-@pytest.mark.parametrize("failure", ["malformed", "unreadable"])
-def test_broken_consent_store_preserves_account_access_and_reports_error(
-    monkeypatch,
-    failure: str,
-) -> None:
-    connection = service(verified)
-    connection.credentials.replace(
-        "india_prod",
-        "key",
-        "secret",
-        state=CredentialState.VERIFIED,
-    )
-    if failure == "malformed":
-        store.path().with_name("consent.json").write_text("not-json")
-    else:
-
-        def unreadable(*args, **kwargs):
-            raise ConsentStorageError("cannot read consent metadata")
-
-        monkeypatch.setattr(connection.consent, "status", unreadable)
-
-    access = asyncio.run(connection.access_state(context("Codex")))
-    status = connection.status(context("Codex"))
-    reads: list[tuple[str, bool]] = []
-
-    async def get(path: str, params=None, *, auth: bool = False):
-        reads.append((path, auth))
-        return {"success": True, "result": {"id": 42}}
-
-    monkeypatch.setattr(connection.client, "get", get)
-    app = build_server(connection_service=connection)
-    account_result = asyncio.run(
-        app.call_tool("get_wallet_balances", {}, context("Codex"))
-    )
-
-    assert access.credentials_ready is True
-    assert access.trading_enabled is False
-    assert access.final_trading_check() is False
-    assert status["credentials_configured"] is True
-    assert status["account_tools_available"] is True
-    assert status["trading"]["enabled"] is False
-    assert status["connection_error"] == "consent_store_unavailable"
-    assert status["consent_error"] == "consent_store_unavailable"
-    assert account_result.is_error is False
-    assert reads == [("/wallet/balances", True)]
-
-
-def test_consent_write_failure_survives_reads_until_a_write_succeeds(
-    monkeypatch,
-) -> None:
-    connection = service(verified)
-    connection.credentials.replace("india_prod", "key", "secret")
-    original_enable = connection.consent.enable
-
-    def unavailable(*args, **kwargs):
-        raise ConsentStorageError("consent metadata is read-only")
-
-    monkeypatch.setattr(connection.consent, "enable", unavailable)
-    failed = action(
-        connection,
-        "Codex",
-        "consent",
-        {
-            "environment": "india_prod",
-            "enabled": True,
-            "acknowledged": True,
-        },
-    )
-
-    assert failed.content["status"] == "rejected"
-    assert connection.consent_error == "consent_store_unavailable"
-    assert connection.status(context("Codex"))["consent_error"] == (
-        "consent_store_unavailable"
-    )
-
-    monkeypatch.setattr(connection.consent, "enable", original_enable)
-    enabled = action(
-        connection,
-        "Codex",
-        "consent",
-        {
-            "environment": "india_prod",
-            "enabled": True,
-            "acknowledged": True,
-        },
-    )
-
-    assert enabled.content["status"] == "enabled"
-    assert connection.status(context("Codex"))["consent_error"] == ""
-
-
-@pytest.mark.parametrize(
-    ("failed_client", "other_client"),
-    [("Codex", ""), ("", "Codex")],
-    ids=[
-        "persistent-failure-survives-memory-write",
-        "memory-failure-survives-persistent-write",
-    ],
-)
-def test_consent_write_recovery_is_scoped_to_the_selected_backend(
-    monkeypatch,
-    failed_client: str,
-    other_client: str,
-) -> None:
-    connection = service(verified)
-    connection.credentials.replace("india_prod", "key", "secret")
-    original_enable = connection.consent.enable
-    failed_persistent = bool(failed_client)
-
-    def fail_selected_backend(binding, **kwargs):
-        if binding.persistent is failed_persistent:
-            raise ConsentStorageError("selected consent backend is read-only")
-        return original_enable(binding, **kwargs)
-
-    monkeypatch.setattr(connection.consent, "enable", fail_selected_backend)
-    arguments = {
-        "environment": "india_prod",
-        "enabled": True,
-        "acknowledged": True,
-    }
-
-    failed = action(connection, failed_client, "consent", arguments)
-    recovered_other = action(connection, other_client, "consent", arguments)
-
-    assert failed.content["status"] == "rejected"
-    assert recovered_other.content["status"] == "enabled"
-    assert connection.status(context(failed_client))["consent_error"] == (
-        "consent_store_unavailable"
-    )
-
-    monkeypatch.setattr(connection.consent, "enable", original_enable)
-    recovered_failed = action(connection, failed_client, "consent", arguments)
-
-    assert recovered_failed.content["status"] == "enabled"
-    assert connection.status(context(failed_client))["consent_error"] == ""
-
-
-@pytest.mark.parametrize(
-    ("failed_client", "other_client"),
-    [("Codex", ""), ("", "Codex")],
-    ids=["persistent", "memory"],
-)
-def test_failed_disable_blocks_trading_but_preserves_reads(
-    monkeypatch,
-    failed_client: str,
-    other_client: str,
-) -> None:
-    connection = service(verified)
-    connection.credentials.replace("india_prod", "key", "secret")
-    arguments = {
-        "environment": "india_prod",
-        "enabled": True,
-        "acknowledged": True,
-    }
-    enabled = action(connection, failed_client, "consent", arguments)
-    captured = asyncio.run(connection.access_state(context(failed_client)))
-    original_disable = connection.consent.disable
-    failed_persistent = bool(failed_client)
-
-    def fail_selected_backend(binding, **kwargs):
-        if binding.persistent is failed_persistent:
-            raise ConsentStorageError("selected consent backend is read-only")
-        return original_disable(binding, **kwargs)
-
-    monkeypatch.setattr(connection.consent, "disable", fail_selected_backend)
-    failed = action(
-        connection,
-        failed_client,
-        "consent",
-        {**arguments, "enabled": False},
-    )
-    healthy_other = action(connection, other_client, "consent", arguments)
-    current = asyncio.run(connection.access_state(context(failed_client)))
-    other = asyncio.run(connection.access_state(context(other_client)))
-    status = connection.status(context(failed_client))
-
-    assert enabled.content["status"] == "enabled"
-    assert captured.trading_enabled is True
-    assert failed.content["status"] == "rejected"
-    assert healthy_other.content["status"] == "enabled"
-    assert current.credentials_ready is True
-    assert current.trading_enabled is False
-    assert current.final_trading_check() is False
-    assert other.trading_enabled is True
-    assert status["trading"]["enabled"] is False
-    assert status["consent_error"] == "consent_store_unavailable"
-
-    _assert_place_order_blocked(
-        monkeypatch,
-        connection,
-        captured.final_trading_check,
-    )
-
-    reads: list[tuple[str, bool]] = []
-
-    async def get(path: str, params=None, *, auth: bool = False):
-        del params
-        reads.append((path, auth))
-        return {"success": True, "result": {"id": 42}}
-
-    monkeypatch.setattr(connection.client, "get", get)
-    app = build_server(connection_service=connection)
-
-    async def read_tools() -> tuple[object, object]:
-        try:
-            public = await app.call_tool(
-                "get_ticker", {"symbol": "BTCUSD"}, context(failed_client)
-            )
-            account = await app.call_tool(
-                "get_wallet_balances", {}, context(failed_client)
-            )
-            return public, account
-        finally:
-            await app.close_live_client()
-
-    public, account = asyncio.run(read_tools())
-
-    assert public.is_error is False
-    assert account.is_error is False
-    assert reads == [("/tickers/BTCUSD", False), ("/wallet/balances", True)]
-
-
-@pytest.mark.parametrize(
-    "persistent",
-    [True, False],
-    ids=["persistent-backend", "memory-backend"],
-)
-def test_failed_disable_stays_denied_after_another_binding_recovers_backend(
-    monkeypatch,
-    persistent: bool,
-) -> None:
-    connection = service(verified, persistent=persistent)
-    connection.credentials.replace("india_prod", "key", "secret")
-    arguments = {
-        "environment": "india_prod",
-        "enabled": True,
-        "acknowledged": True,
-    }
-    action(connection, "Codex", "consent", arguments)
-    action(connection, "Claude", "consent", arguments)
-    captured = asyncio.run(connection.access_state(context("Codex")))
-    original_disable = connection.consent.disable
-
-    def fail_codex(binding, **kwargs):
-        if binding.client_name == "Codex":
-            raise ConsentStorageError("Codex consent backend is read-only")
-        return original_disable(binding, **kwargs)
-
-    monkeypatch.setattr(connection.consent, "disable", fail_codex)
-    failed = action(
-        connection,
-        "Codex",
-        "consent",
-        {**arguments, "enabled": False},
-    )
-    monkeypatch.setattr(connection.consent, "disable", original_disable)
-    other = action(
-        connection,
-        "Claude",
-        "consent",
-        {**arguments, "enabled": False},
-    )
-
-    current = asyncio.run(connection.access_state(context("Codex")))
-    status = connection.status(context("Codex"))
-    assert failed.content["status"] == "rejected"
-    assert other.content["status"] == "disabled"
-    assert captured.final_trading_check() is False
-    assert current.trading_enabled is False
-    assert status["trading"]["enabled"] is False
-    assert status["consent_error"] == "consent_store_unavailable"
-
-    _assert_place_order_blocked(
-        monkeypatch,
-        connection,
-        captured.final_trading_check,
-    )
-
-    recovered = action(
-        connection,
-        "Codex",
-        "consent",
-        {**arguments, "enabled": False},
-    )
-    reenabled = action(connection, "Codex", "consent", arguments)
-    allowed = asyncio.run(connection.access_state(context("Codex")))
-
-    assert recovered.content["status"] == "disabled"
-    assert reenabled.content["status"] == "enabled"
-    assert allowed.trading_enabled is True
-    assert allowed.final_trading_check() is True
-    assert connection.status(context("Codex"))["consent_error"] == ""
-
-
-@pytest.mark.parametrize(
-    "persistent",
-    [True, False],
-    ids=["persistent-backend", "memory-backend"],
-)
-def test_credential_rotation_drops_an_obsolete_binding_denial(
-    monkeypatch,
-    persistent: bool,
-) -> None:
-    connection = service(verified, persistent=persistent)
-    connection.credentials.replace("india_prod", "old-key", "old-secret")
-    arguments = {
-        "environment": "india_prod",
-        "enabled": True,
-        "acknowledged": True,
-    }
-    original_enable = connection.consent.enable
-
-    def fail_enable(*args, **kwargs):
-        raise ConsentStorageError("consent backend is read-only")
-
-    monkeypatch.setattr(connection.consent, "enable", fail_enable)
-    failed = action(connection, "Codex", "consent", arguments)
-    monkeypatch.setattr(connection.consent, "enable", original_enable)
-    rotated = action(
-        connection,
-        "Codex",
-        "credentials",
-        {
-            "operation": "replace",
-            "environment": "india_prod",
-            "api_key": "new-key",
-            "api_secret": "new-secret",
-        },
-    )
-    enabled = action(
-        connection,
-        "Codex",
-        "consent",
-        arguments,
-        rotated.revision,
-    )
-
-    assert failed.content["status"] == "rejected"
-    assert rotated.content["status"] == "saved"
-    assert enabled.content["status"] == "enabled"
-    assert connection.status(context("Codex"))["consent_error"] == ""
-
-
-def test_failed_environment_revocation_survives_backend_recovery(
-    monkeypatch,
-) -> None:
-    connection = service(verified)
-    connection.credentials.replace("india_prod", "prod-key", "prod-secret")
-    connection.credentials.replace("india_testnet", "test-key", "test-secret")
-    arguments = {
-        "environment": "india_prod",
-        "enabled": True,
-        "acknowledged": True,
-    }
-    action(connection, "Codex", "consent", arguments)
-    action(connection, "Claude", "consent", arguments)
-    captured = asyncio.run(connection.access_state(context("Codex")))
-    original_revoke = connection.consent.revoke_environment
-
-    def fail_prod(environment: str) -> frozenset[ConsentBackend]:
-        if environment == "india_prod":
-            raise ConsentRevocationError(
-                "persistent consent metadata is read-only",
-                failed_backend=ConsentBackend.PERSISTENT,
-                written=frozenset(),
-                checked=frozenset({ConsentBackend.MEMORY}),
-            )
-        return original_revoke(environment)
-
-    monkeypatch.setattr(connection.consent, "revoke_environment", fail_prod)
-    failed = action(
-        connection,
-        "Codex",
-        "credentials",
-        {"operation": "activate", "environment": "india_testnet"},
-    )
-    monkeypatch.setattr(connection.consent, "revoke_environment", original_revoke)
-    recovered_backend = action(
-        connection,
-        "Claude",
-        "consent",
-        {**arguments, "enabled": False},
-    )
-
-    current = asyncio.run(connection.access_state(context("Codex")))
-    assert failed.content["status"] == "rejected"
-    assert recovered_backend.content["status"] == "disabled"
-    assert captured.final_trading_check() is False
-    assert current.trading_enabled is False
-    assert connection.status(context("Codex"))["consent_error"] == (
-        "consent_store_unavailable"
-    )
-
-    _assert_place_order_blocked(
-        monkeypatch,
-        connection,
-        captured.final_trading_check,
-    )
-
-    recovered_scope = action(
-        connection,
-        "Codex",
-        "credentials",
-        {"operation": "activate", "environment": "india_testnet"},
-    )
-
-    assert recovered_scope.content["status"] == "selected"
-    assert connection.status(context("Codex"))["consent_error"] == ""
-
-
-def test_environment_revocation_expires_for_an_external_credential_successor(
-    monkeypatch,
-) -> None:
-    connection = service(verified)
-    connection.credentials.replace("india_prod", "old-key", "old-secret")
-    connection.credentials.replace("india_testnet", "test-key", "test-secret")
-    arguments = {
-        "environment": "india_prod",
-        "enabled": True,
-        "acknowledged": True,
-    }
-    action(connection, "Codex", "consent", arguments)
-    action(connection, "Claude", "consent", arguments)
-    captured = asyncio.run(connection.access_state(context("Codex")))
-    original_environment_revoke = connection.consent.revoke_environment
-
-    def fail_environment(environment: str) -> frozenset[ConsentBackend]:
-        if environment == "india_prod":
-            raise ConsentRevocationError(
-                "persistent consent metadata is read-only",
-                failed_backend=ConsentBackend.PERSISTENT,
-                written=frozenset(),
-                checked=frozenset({ConsentBackend.MEMORY}),
-            )
-        return original_environment_revoke(environment)
-
-    monkeypatch.setattr(
-        connection.consent,
-        "revoke_environment",
-        fail_environment,
-    )
-    failed = action(
-        connection,
-        "Codex",
-        "credentials",
-        {"operation": "activate", "environment": "india_testnet"},
-    )
-    monkeypatch.setattr(
-        connection.consent,
-        "revoke_environment",
-        original_environment_revoke,
-    )
-    recovered_backend = action(
-        connection,
-        "Claude",
-        "consent",
-        {**arguments, "enabled": False},
-    )
-    same_identity = asyncio.run(connection.access_state(context("Codex")))
-
-    assert failed.content["status"] == "rejected"
-    assert recovered_backend.content["status"] == "disabled"
-    assert same_identity.trading_enabled is False
-    assert captured.final_trading_check() is False
-    assert store.environment_state("india_prod") == ("india_prod", 0)
-
-    connection.credentials.replace("india_prod", "new-key", "new-secret")
-    original_identity_revoke = connection.consent.revoke_identity
-
-    def fail_identity(binding: ConsentBinding) -> frozenset[ConsentBackend]:
-        del binding
-        raise ConsentRevocationError(
-            "persistent consent metadata is read-only",
-            failed_backend=ConsentBackend.PERSISTENT,
-            written=frozenset(),
-            checked=frozenset({ConsentBackend.MEMORY}),
-        )
-
-    monkeypatch.setattr(connection.consent, "revoke_identity", fail_identity)
-    rotated = connection.status(context("Codex"))
-    monkeypatch.setattr(
-        connection.consent,
-        "revoke_identity",
-        original_identity_revoke,
-    )
-    enabled = action(connection, "Codex", "consent", arguments)
-    successor = asyncio.run(connection.access_state(context("Codex")))
-
-    assert rotated["trading"]["enabled"] is False
-    assert rotated["consent_error"] == "consent_store_unavailable"
-    assert enabled.content["status"] == "enabled"
-    assert successor.trading_enabled is True
-    assert successor.final_trading_check() is True
-    assert captured.final_trading_check() is False
-    assert connection.status(context("Codex"))["consent_error"] == ""
-    assert store.environment_state("india_prod") == ("india_prod", 0)
-
-
-def test_environment_revocation_expires_for_the_first_external_credential(
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    connection = service(verified)
-    connection.credentials.replace("india_testnet", "test-key", "test-secret")
-    original_environment_revoke = connection.consent.revoke_environment
-
-    def fail_prod(environment: str) -> frozenset[ConsentBackend]:
-        if environment == "india_prod":
-            raise ConsentRevocationError(
-                "persistent consent metadata is read-only",
-                failed_backend=ConsentBackend.PERSISTENT,
-                written=frozenset(),
-                checked=frozenset({ConsentBackend.MEMORY}),
-            )
-        return original_environment_revoke(environment)
-
-    monkeypatch.setattr(connection.consent, "revoke_environment", fail_prod)
-    failed = action(
-        connection,
-        "Codex",
-        "credentials",
-        {"operation": "activate", "environment": "india_testnet"},
-    )
-    monkeypatch.setattr(
-        connection.consent,
-        "revoke_environment",
-        original_environment_revoke,
-    )
-    credential = connection.credentials.replace(
-        "india_prod",
-        "new-key",
-        "new-secret",
-    )
-    arguments = {
-        "environment": "india_prod",
-        "enabled": True,
-        "acknowledged": True,
-    }
-    enabled = action(connection, "Codex", "consent", arguments)
-    current = asyncio.run(connection.access_state(context("Codex")))
-
-    assert failed.content["status"] == "rejected"
-    assert credential.generation == 1
-    assert enabled.content["status"] == "enabled"
-    assert current.trading_enabled is True
-    assert current.final_trading_check() is True
-    assert connection.status(context("Codex"))["consent_error"] == ""
-    assert store.environment_state("india_prod") == ("india_prod", 0)
-
-
-def test_successful_empty_environment_revocation_clears_failed_scope(
-    monkeypatch,
-) -> None:
-    connection = service(verified)
-    connection.credentials.replace("india_prod", "prod-key", "prod-secret")
-    connection.credentials.replace("india_testnet", "test-key", "test-secret")
-    original_revoke = connection.consent.revoke_environment
-
-    def fail_testnet(environment: str) -> frozenset[ConsentBackend]:
-        if environment == "india_testnet":
-            raise ConsentRevocationError(
-                "persistent consent metadata is read-only",
-                failed_backend=ConsentBackend.PERSISTENT,
-                written=frozenset(),
-                checked=frozenset({ConsentBackend.MEMORY}),
-            )
-        return original_revoke(environment)
-
-    monkeypatch.setattr(connection.consent, "revoke_environment", fail_testnet)
-    failed = action(
-        connection,
-        "Codex",
-        "credentials",
-        {"operation": "activate", "environment": "india_testnet"},
-    )
-    monkeypatch.setattr(connection.consent, "revoke_environment", original_revoke)
-    enabled = action(
-        connection,
-        "Codex",
-        "consent",
-        {
-            "environment": "india_prod",
-            "enabled": True,
-            "acknowledged": True,
-        },
-    )
-
-    assert failed.content["status"] == "rejected"
-    assert enabled.content["status"] == "enabled"
-    assert connection.status(context("Codex"))["consent_error"] == (
-        "consent_store_unavailable"
-    )
-
-    recovered = action(
-        connection,
-        "Codex",
-        "credentials",
-        {"operation": "activate", "environment": "india_testnet"},
-    )
-
-    assert recovered.content["status"] == "selected"
-    assert connection.status(context("Codex"))["consent_error"] == ""
-
-
-def test_environment_generation_drops_denials_after_active_binding_was_lost(
-    monkeypatch,
-) -> None:
-    connection = service(verified)
-    connection.credentials.replace("india_prod", "prod-key", "prod-secret")
-    connection.credentials.replace("india_testnet", "test-key", "test-secret")
-    original_enable = connection.consent.enable
-
-    def fail_enable(*args, **kwargs):
-        raise ConsentStorageError("persistent consent metadata is read-only")
-
-    monkeypatch.setattr(connection.consent, "enable", fail_enable)
-    failed = action(
-        connection,
-        "Codex",
-        "consent",
-        {
-            "environment": "india_prod",
-            "enabled": True,
-            "acknowledged": True,
-        },
-    )
-    monkeypatch.setattr(connection.consent, "enable", original_enable)
-    connection._active_binding = None
-
-    assert store.write({"DELTA_MCP_ENV": "india_testnet"}) is None
-    enabled = action(
-        connection,
-        "Claude",
-        "consent",
-        {"environment": "india_testnet", "enabled": True},
-    )
-
-    assert failed.content["status"] == "rejected"
-    assert enabled.content["status"] == "enabled"
-    assert connection.status(context("Claude"))["consent_error"] == ""
-
-
-def test_environment_scope_expires_after_active_binding_was_unavailable(
-    monkeypatch,
-) -> None:
-    connection = service(verified)
-    connection.credentials.replace("india_prod", "prod-key", "prod-secret")
-    connection.credentials.replace("india_testnet", "test-key", "test-secret")
-    original_revoke_environment = connection.consent.revoke_environment
-
-    def fail_prod(environment: str) -> frozenset[ConsentBackend]:
-        if environment == "india_prod":
-            raise ConsentRevocationError(
-                "persistent consent metadata is read-only",
-                failed_backend=ConsentBackend.PERSISTENT,
-                written=frozenset(),
-                checked=frozenset({ConsentBackend.MEMORY}),
-            )
-        return original_revoke_environment(environment)
-
-    monkeypatch.setattr(connection.consent, "revoke_environment", fail_prod)
-    failed = action(
-        connection,
-        "Codex",
-        "credentials",
-        {"operation": "activate", "environment": "india_testnet"},
-    )
-    monkeypatch.setattr(
-        connection.consent,
-        "revoke_environment",
-        original_revoke_environment,
-    )
-    original_resolve = connection.credentials.resolve
-    original_revoke_identity = connection.consent.revoke_identity
-
-    def fail_resolve(*args, **kwargs):
-        raise CredentialStoreError("credential metadata is temporarily unavailable")
-
-    def fail_identity(*args, **kwargs):
-        raise ConsentRevocationError(
-            "persistent consent metadata is read-only",
-            failed_backend=ConsentBackend.PERSISTENT,
-            written=frozenset(),
-            checked=frozenset({ConsentBackend.MEMORY}),
-        )
-
-    monkeypatch.setattr(connection.credentials, "resolve", fail_resolve)
-    monkeypatch.setattr(connection.consent, "revoke_identity", fail_identity)
-    connection._reconcile()
-    monkeypatch.setattr(connection.credentials, "resolve", original_resolve)
-    monkeypatch.setattr(connection.consent, "revoke_identity", original_revoke_identity)
-
-    assert store.write({"DELTA_MCP_ENV": "india_testnet"}) is None
-    enabled = action(
-        connection,
-        "Claude",
-        "consent",
-        {"environment": "india_testnet", "enabled": True},
-    )
-
-    assert failed.content["status"] == "rejected"
-    assert enabled.content["status"] == "enabled"
-    assert connection.status(context("Claude"))["consent_error"] == ""
-
-
-def test_environment_generation_expires_a_failed_disconnect_scope(
-    monkeypatch,
-) -> None:
-    connection = service(verified)
-    connection.credentials.replace("india_prod", "prod-key", "prod-secret")
-    connection.credentials.replace("india_testnet", "test-key", "test-secret")
-    original_revoke = connection.consent.revoke_before
-
-    def fail_revoke(*args, **kwargs):
-        raise ConsentRevocationError(
-            "persistent consent metadata is read-only",
-            failed_backend=ConsentBackend.PERSISTENT,
-            written=frozenset(),
-            checked=frozenset({ConsentBackend.MEMORY}),
-        )
-
-    monkeypatch.setattr(connection.consent, "revoke_before", fail_revoke)
-    disconnected = action(
-        connection,
-        "Codex",
-        "credentials",
-        {"operation": "disconnect", "environment": "india_prod"},
-    )
-    monkeypatch.setattr(connection.consent, "revoke_before", original_revoke)
-
-    assert store.write({"DELTA_MCP_ENV": "india_testnet"}) is None
-    enabled = action(
-        connection,
-        "Claude",
-        "consent",
-        {"environment": "india_testnet", "enabled": True},
-    )
-
-    assert disconnected.content["status"] == "disconnected"
-    assert enabled.content["status"] == "enabled"
-    assert connection.status(context("Claude"))["consent_error"] == ""
 
 
 @pytest.mark.parametrize("disable_fails", [False, True], ids=["disabled", "failed"])
